requirements.txt
# GeoTemporalFusion - Python Dependencies
# ================================================================================
# For Hugging Face Spaces Deployment (CPU Inference)
# ================================================================================

# Core ML Framework (CPU version for free tier)
torch>=2.0.0
torchvision>=0.15.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# API Utilities
pydantic>=2.0.0
python-jose>=3.3.0

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.1.0
